+++ cernml/coi/checkers/__init__.py
@@ -1,5 +1,6 @@
 """Provides the function `check()` to validate a `Problem`."""
 
+from ._configurable import check_configurable
 from ._env import check_env
 from ._full_check import check
 from ._func_opt import check_function_optimizable
@@ -10,6 +11,7 @@ from ._single_opt import check_single_optimizable
 __all__ = [
     "assert_range",
     "check",
+    "check_configurable",
     "check_env",
     "check_function_optimizable",
     "check_problem",